Code cleanup and rewritten build system

+++ SparseGrids/tasgridWrapper.hpp
@@ -183,7 +183,7 @@ protected:
     double* readTransform() const;
     int* readLevelLimits(int num_weights) const;
 
-    static void readMatrix(const char *filename, int &rows, int &cols, double* &mat);
+    static void readMatrix(const char *filename, size_t &rows, size_t &cols, double* &mat);
     static void writeMatrix(const char *filename, int rows, int cols, const double mat[], bool ascii);
     static void printMatrix(int rows, int cols, const double mat[]);
